+++ app/Http/Controllers/AdminController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\actionsAdmin\AdminReportReplyRequest;
 use App\Http\Requests\actionsAdmin\AdminSearchUserRequest;
 use App\Http\Requests\actionsAdmin\AdminUpdateUserRequest;
 use App\Http\Requests\actionsUser\EditProductRequest;
@@ -102,9 +103,35 @@ class AdminController extends Controller
         if ($status != 'closed' && $status != 'opened') {
             return redirect()->back()->with('error', 'Invalid status.');
         }
-
+        $reports = null;
+        if ($status == "opened") {
+            $reports = Report::where("status", 1)->orderBy("created_at", "desc")->get();
+        } else if ($status == "closed") {
+            $reports = Report::where("status", 0)->orderBy("created_at", "desc")->get();
+        }
         $user = $request->user();
-        Report::where("");
-        return view('admin.reports');
+
+        return view('admin.reports')->with([
+            'user' => $user,
+            'reports' => $reports,
+        ]);
+    }
+
+    public function showReport(Request $request, Report $report) {
+        $user = $request->user();
+
+        return view('admin.report')->with([
+            'user' => $user,
+            'report' => $report,
+        ]);
+    }
+
+    public function replyReport(AdminReportReplyRequest $request, Report $report) {
+        $user = $request->user();
+        $validated_data = $request->validated();
+
+        $this->adminService->replyReport($user, $report, $validated_data);
+
+        return back()->with('success', 'You have successfully replied to report.');
     }
 }